post_train_infra/docker/rag-orchestrator/requirements.txt
# RAG orchestrator dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
weaviate-client==3.26.2
pydantic>=2.5.0
